08/03/2018 1142
+1 Warp is now available, although it's kind of bad. But still available.

+++ Assets/Q9Core/ScaleSpace.cs
@@ -37,7 +37,7 @@ namespace Q9Core
 
         }
 
-        public static void Warp(DoubleVector3 d, float s)
+        public static void Warp(DoubleVector3 d, double s)
         {
             SetApparentPosition(DoubleVector3.MoveTowards(apparentPosition, d, s * 149597870700));
         }
